+++ test/test.spec.ts
@@ -6,16 +6,14 @@ import chaiAsPromised from 'chai-as-promised';
 import puppeteer from 'puppeteer';
 
 import { Dappeteer, RECOMMENDED_METAMASK_VERSION } from '../src';
-import { clickOnLogo, openProfileDropdown } from '../src/helpers';
 import * as dappeteer from '../src/index';
 
 import deploy from './deploy';
+import { addNetworkTests } from './tests/addNetwork';
+import { importPKTests } from './tests/importPK';
+import { pause } from './utils';
 
 chaiUse(chaiAsPromised);
-
-function pause(seconds: number): Promise<void> {
-  return new Promise((res) => setTimeout(res, 1000 * seconds));
-}
 
 function getCounterNumber(contract): Promise<number> {
   return contract.methods
@@ -33,7 +31,7 @@ async function clickElement(page, selector): Promise<void> {
   await element.click();
 }
 
-let testContract, browser, metamask: Dappeteer, testPage;
+export let testContract, browser, metamask: Dappeteer, testPage;
 
 describe('dappeteer', () => {
   before(async () => {
@@ -55,6 +53,7 @@ describe('dappeteer', () => {
     console.log(`::set-output name=version::${files.pop().replace(/_/g, '.')}`);
   });
 
+  // validate dappateer setup
   it('should be deployed, contract', async () => {
     expect(testContract).to.be.ok;
     expect(testContract.address).to.be.ok;
@@ -74,88 +73,10 @@ describe('dappeteer', () => {
     expect(await testPage.title()).to.be.equal('Local metamask test');
   });
 
-  describe('test addNetwork method', async () => {
-    after(async () => {
-      await metamask.switchNetwork('local');
-      await metamask.helpers.deleteNetwork('Binance Smart Chain');
-      await pause(0.5);
-      await metamask.helpers.deleteNetwork('Optimistic Ethereum Testnet Kovan');
-      await pause(0.5);
-      await metamask.helpers.deleteNetwork('KCC Testnet');
-      await pause(0.5);
-    });
+  describe('test addNetwork method', addNetworkTests.bind(this));
+  describe('test importPK method', importPKTests.bind(this));
 
-    it('should add network with required params and symbol and explorer', async () => {
-      await metamask.addNetwork({
-        networkName: 'Binance Smart Chain',
-        rpc: 'https://data-seed-prebsc-1-s1.binance.org:8545/',
-        chainId: 97,
-        symbol: 'BNB',
-        explorer: 'https://testnet.bscscan.com',
-      });
-
-      const selectedNetwork = await metamask.page.evaluate(
-        () => (document.querySelector('.network-display > span:nth-child(2)') as HTMLSpanElement).innerHTML,
-      );
-      expect(selectedNetwork).to.be.equal('Binance Smart Chain');
-    });
-
-    it('should fail to add already added network', async () => {
-      await expect(
-        metamask.addNetwork({
-          networkName: 'Binance Smart Chain',
-          rpc: 'https://data-seed-prebsc-1-s1.binance.org:8545/',
-          chainId: 97,
-          symbol: 'BNB',
-          explorer: 'https://testnet.bscscan.com',
-        }),
-      ).to.be.rejectedWith(SyntaxError);
-
-      await clickOnLogo(metamask.page);
-    });
-
-    it('should fail to add network with wrong chain ID', async () => {
-      await expect(
-        metamask.addNetwork({
-          networkName: 'Optimistic Ethereum Testnet Kovan',
-          rpc: 'https://kovan.optimism.io/',
-          chainId: 420,
-          explorer: 'https://kovan-optimistic.etherscan.io',
-        }),
-      ).to.be.rejectedWith(SyntaxError);
-
-      await clickOnLogo(metamask.page);
-    });
-
-    it('should add network with explorer', async () => {
-      await metamask.addNetwork({
-        networkName: 'Optimistic Ethereum Testnet Kovan',
-        rpc: 'https://kovan.optimism.io/',
-        chainId: 69,
-        explorer: 'https://kovan-optimistic.etherscan.io',
-      });
-
-      const selectedNetwork = await metamask.page.evaluate(
-        () => (document.querySelector('.network-display > span:nth-child(2)') as HTMLSpanElement).innerHTML,
-      );
-      expect(selectedNetwork).to.be.equal('Optimistic Ethereum Testnet Kovan');
-    });
-
-    it('should add network with symbol', async () => {
-      await metamask.addNetwork({
-        networkName: 'KCC Testnet',
-        rpc: 'https://rpc-testnet.kcc.network',
-        chainId: 322,
-        symbol: 'fejk',
-      });
-
-      const selectedNetwork = await metamask.page.evaluate(
-        () => (document.querySelector('.network-display > span:nth-child(2)') as HTMLSpanElement).innerHTML,
-      );
-      expect(selectedNetwork).to.be.equal('KCC Testnet');
-    });
-  });
-
+  // TODO: add more cases
   it('should switch network, localhost', async () => {
     await metamask.switchNetwork('localhost');
 
@@ -163,51 +84,6 @@ describe('dappeteer', () => {
       () => (document.querySelector('.network-display > span:nth-child(2)') as HTMLSpanElement).innerHTML,
     );
     expect(selectedNetwork).to.be.equal('Localhost 8545');
-  });
-
-  describe('test importPK method', async () => {
-    afterEach(async () => {
-      await clickOnLogo(metamask.page);
-    });
-
-    after(async () => {
-      await metamask.helpers.deleteAccount(2);
-      await pause(0.5);
-    });
-
-    it('should import private key', async () => {
-      const countAccounts = async (): Promise<number> => {
-        await openProfileDropdown(metamask.page);
-        const container = await metamask.page.$('.account-menu__accounts');
-        const count = (await container.$$('.account-menu__account')).length;
-        await openProfileDropdown(metamask.page);
-        return count;
-      };
-
-      const beforeImport = await countAccounts();
-      await metamask.importPK('4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b10');
-      const afterImport = await countAccounts();
-
-      expect(beforeImport + 1).to.be.equal(afterImport);
-    });
-
-    it('should throw error on duplicated private key', async () => {
-      await expect(
-        metamask.importPK('4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b10'),
-      ).to.be.rejectedWith(SyntaxError);
-    });
-
-    it('should throw error on wrong key', async () => {
-      await expect(
-        metamask.importPK('4f3edf983ac636a65a$@!ce7c78d9aa706d3b113bce9c46f30d7d21715b23b10'),
-      ).to.be.rejectedWith(SyntaxError);
-    });
-
-    it('should throw error on to short key', async () => {
-      await expect(
-        metamask.importPK('4f3edf983ac636a65ace7c78d9aa706d3b113bce9c46f30d7d21715b23b10'),
-      ).to.be.rejectedWith(SyntaxError);
-    });
   });
 
   describe('test switchAccount method', async () => {
